Don't build splitToUDPAddr on windows. It's not used.
