+++ pkg/client/daemon/outbound_unix.go
@@ -1,0 +1,19 @@
+// +build !windows
+
+package daemon
+
+import (
+	"net"
+
+	"github.com/telepresenceio/telepresence/v2/pkg/iputil"
+)
+
+// splitToUDPAddr splits the given address into an UDPAddr. It's
+// an  error if the address is based on a hostname rather than an IP.
+func splitToUDPAddr(netAddr net.Addr) (*net.UDPAddr, error) {
+	ip, port, err := iputil.SplitToIPPort(netAddr)
+	if err != nil {
+		return nil, err
+	}
+	return &net.UDPAddr{IP: ip, Port: int(port)}, nil
+}
